add the get available promo codes and api's and testcases

+++ orders/interactors/dtos.py
@@ -23,8 +23,8 @@ class PromoCodeDTO:
     promo_code_id: int
     code: str
     discount_type: str
-    discount_value: float
-    min_order_value: float
+    discount_value: Decimal
+    min_order_value: Decimal
     max_usage: int
     valid_from: datetime
     valid_until: datetime
